+++ packages/expo/types/test/fixtures/exposed-types.ts
@@ -1,8 +1,8 @@
 import { Bugsnag } from "../../..";
 let bugsnagInstance: Bugsnag.Client | undefined = undefined;
-export function notify(error: Bugsnag.NotifiableError, opts?: Bugsnag.NotifyOpts): void {
+export function notify(error: Bugsnag.NotifiableError, onError?: Bugsnag.OnError): void {
   if (bugsnagInstance === undefined) {
     return
   }
-  return bugsnagInstance.notify(error, opts)
+  return bugsnagInstance.notify(error, onError)
 }