+++ packages/plugin-express/src/express.js
@@ -32,7 +32,7 @@ module.exports = {
 
       // unhandled errors caused by this request
       dom.on('error', (err) => {
-        req.bugsnag.notify(createEventFromErr(err, handledState), {}, (e, event) => {
+        req.bugsnag.notify(createEventFromErr(err, handledState), () => {}, (e, event) => {
           if (e) client._logger.error('Failed to send event to Bugsnag')
           req.bugsnag.config.onUncaughtException(err, event, client._logger)
         })
@@ -52,7 +52,11 @@ module.exports = {
         client._logger.warn(
           'req.bugsnag is not defined. Make sure the @bugsnag/plugin-express requestHandler middleware is added first.'
         )
-        client.notify(createEventFromErr(err, handledState, getRequestAndMetaDataFromReq(req)))
+        client.notify(createEventFromErr(err, handledState), (event) => {
+          const { metaData, request } = getRequestAndMetaDataFromReq(req)
+          event.request = { ...request }
+          event.metaData = { ...metaData }
+        })
       }
       next(err)
     }